backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
sqlalchemy>=2.0.23
pydantic>=2.0.0
psycopg2-binary
python-dotenv>=1.0.0